Add file for app component
