+++ src/components/Todoinput.jsx
@@ -2,10 +2,10 @@ import React from 'react';
 
 function Todoinput({ inputText, onClick, onChange }) {
   return (
-    <form>
+    <span>
       <input type="text" value={inputText} onChange={onChange} />
       <button type="button" onClick={() => onClick(inputText)}>추가</button>
-    </form>
+    </span>
   );
 }
